feat: portable u32 PCG-hash RNG for CPU/GPU parity
Replace rand_pcg::Pcg32 (u64 state) with a counter-based PCG hash over
pure u32 ops (Jarzynski & Olano 2020). Same (key, seed) produces the same
sequence on CPU and will reproduce bit-for-bit in WGSL. Frozen draw vector
[0xa1dd5847, 0x13248a2d, 0x37278def, 0xde1d5b62] is the GPU parity anchor.
Key derivation in sample_pixel switched to mix2(pixel, sample_idx); all
seed/key parameters narrowed from u64 to u32 across tracer, lighttrace,
and volume.

+++ spectral-core/src/lighttrace.rs
@@ -154,9 +154,9 @@ pub fn trace_photon(
 
 /// Emit `n` photons from the beam (wavelengths uniform over the visible band) and
 /// splat survivors onto the screen.
-pub fn trace(scene: &Scene, screen: &mut Screen, beam: &Beam, n: u32, seed: u64) {
+pub fn trace(scene: &Scene, screen: &mut Screen, beam: &Beam, n: u32, seed: u32) {
     for i in 0..n {
-        let mut rng = PathRng::new(i as u64, seed);
+        let mut rng = PathRng::new(i, seed);
         let origin = beam.corner + beam.u * rng.next_f32() + beam.v * rng.next_f32();
         let lambda = LAMBDA_MIN + LAMBDA_RANGE * rng.next_f32();
         if let Some((s, r, power)) = trace_photon(scene, screen, Ray { origin, dir: beam.dir }, lambda, &mut rng) {
@@ -176,7 +176,7 @@ mod tests {
     fn mean_landing(scene: &Scene, screen: &Screen, beam: &Beam, lambda: f32, n: u32) -> Option<(f32, f32)> {
         let (mut sx, mut sr, mut cnt) = (0.0f32, 0.0f32, 0u32);
         for i in 0..n {
-            let mut rng = PathRng::new(i as u64, 1);
+            let mut rng = PathRng::new(i, 1);
             let origin = beam.corner + beam.u * rng.next_f32() + beam.v * rng.next_f32();
             if let Some((s, r, _)) = trace_photon(scene, screen, Ray { origin, dir: beam.dir }, lambda, &mut rng) {
                 sx += s; sr += r; cnt += 1;
